수정내용
- 기기없는 그룹 삭제 함수 중복Idx값이 없을 경우 예외처리

+++ src/main/java/com/oauth/utils/Common.java
@@ -373,9 +373,10 @@ public class Common {
                 .collect(Collectors.toList());
 
         log.info("inviteIdxList에서 registIdxList에 없는 값: " + uniqueInviteIdxList);
-
-        int result = deviceMapper.deleteNoDeviceGroupByList(uniqueInviteIdxList);
-        log.info("DELETE GROUP RESULT: " + result);
+        if(!uniqueInviteIdxList.isEmpty()){
+            int result = deviceMapper.deleteNoDeviceGroupByList(uniqueInviteIdxList);
+            log.info("DELETE GROUP RESULT: " + result);
+        }
     }
 
 }
